hw8/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
numpy==1.26.4
tensorflow-cpu==2.15.0
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
imbalanced-learn==0.12.2